@@ -3,39 +3,44 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev --turbopack",
     "build": "next build",
     "start": "next start",
     "lint": "next lint"
   },
   "dependencies": {
-    "@mdx-js/loader": "^2.3.0",
-    "@mdx-js/react": "^2.3.0",
-    "@next/mdx": "^13.4.19",
-    "@react-three/drei": "^9.80.9",
-    "@react-three/fiber": "^8.13.7",
-    "@tailwindcss/typography": "^0.5.9",
-    "@types/mdx": "^2.0.7",
-    "@types/node": "20.5.9",
-    "@types/react": "18.2.21",
-    "@types/react-dom": "18.2.7",
-    "@types/three": "^0.155.1",
-    "autoprefixer": "10.4.15",
-    "eslint": "8.48.0",
-    "eslint-config-next": "13.4.19",
-    "framer-motion": "^10.16.2",
+    "@mdx-js/loader": "^3.1.0",
+    "@mdx-js/react": "^3.1.0",
+    "@next/mdx": "15.3.2",
+    "@react-three/drei": "^10.0.8",
+    "@react-three/fiber": "^9.1.2",
+    "@tailwindcss/postcss": "^4.1.7",
+    "@tailwindcss/typography": "^0.5.16",
+    "@types/mdx": "^2.0.13",
+    "@types/node": "22.15.21",
+    "@types/react": "19.1.5",
+    "@types/react-dom": "19.1.5",
+    "@types/three": "^0.176.0",
+    "autoprefixer": "^10.4.17",
+    "eslint": "9.27.0",
+    "eslint-config-next": "15.3.2",
     "lodash.throttle": "^4.1.1",
-    "next": "13.4.19",
-    "next-themes": "^0.2.1",
-    "postcss": "8.4.29",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "tailwindcss": "3.3.3",
-    "three": "^0.156.0",
-    "typescript": "5.2.2"
+    "motion": "^12.12.1",
+    "next": "15.3.2",
+    "next-themes": "^0.4.6",
+    "postcss": "^8.4.35",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
+    "tailwindcss": "^3.4.1",
+    "three": "^0.176.0",
+    "typescript": "5.8.3"
   },
   "devDependencies": {
-    "@types/lodash.throttle": "^4.1.7",
-    "debug": "^4.3.4"
+    "@types/lodash.throttle": "^4.1.9",
+    "debug": "^4.4.1"
+  },
+  "overrides": {
+    "@types/react": "19.1.5",
+    "@types/react-dom": "19.1.5"
   }
 }
